-Codechange: Drop DirectoryReader use in ConfigFile::LoadFromDirectoryList. (PovAddictW)
It was using a DirectoryReader but not reading any directory, just using
its methods to concatenate a filename to a directory path and to check if
the final path was a file, which didn't make much sense.

Instead, add new functions in fileio separate from DirectoryReader
to test if a path is a file or a directory.



git-svn-id: https://freerct.googlecode.com/svn/trunk@1122 d3bf82d2-0c16-e458-5408-27be57bd1276